The search / filter box should have a label.
Makes it more accessible for keyboard users, and for people who use a mouse adds the ability to click the text next to the field to focus the element - which is useful.

A label should offer just as much styling opportunity as a div. If this isn't an acceptable soultion, one could create a new label element within the exisiting div, and make it point to the input.