Added functionality to insert html elements and nodes into td (using mRender).
A developer may way want to append an element or node in cell data rather than only html. This is particularily useful when attempting to retain binding connections between the dom and an object. For example, useful to bind a Backbone View object to a dom node; otherwise, without this, using to html() on the backbone view destroys the binding connection.

The two new functions used to test isElement or isNode were both taken from StackOverflow http://stackoverflow.com/questions/384286/javascript-isdom-how-do-you-check-if-a-javascript-object-is-a-dom-object.