do "zeroOut" after collecting column width.
Removing border-top/bottom or padding-top/bottom will change height of the table.
If the table size was changed smaller than divBody, the vertical scrollbar will disappear, and column width may change wider.
It is not the expected result because height of divBody has already been expanded including borders and paddings of the header.