Bug Fix events on inner elements overriden by the mere existence of a custom render function
This addresses the following scenario:

I am adding existing (non-rendered to the DOM) TRs to DT with events
attached to their children.

I want a custom render function for providing filter and sort data, but
not display data. If there is a custom render function the cell data is
overwritten and event are lost.

This could be prevented if the render function can signal DT that it
doesn’t want to override the data and it wants to use the default
display data. Returning `undefined` can be a good signal.

To address this issue, before overwriting the innerData of a cell using
the display data from the custom render function you can check whether
the data is `null` (`undefined` gets turned into `null` at this point).

And for all intents and purposes there is no point in overwriting the
cell content with `null` or `undefined`. Of course overwriting it with
an empty string (`””`) works fine.